Pulling settings from environment variables
Hard-coding the settings in the image is not the way to go. Doing so makes it very difficult to use the image in varying environments where we cannot control certain values (esp. hostnames). Using environment variables allow us to easily create new containers with different settings, and not have to constantly rebuild images.